removed the intranet random enum

+++ Application/Data/Stores/IntranetCmdStore.cpp
@@ -9,8 +9,8 @@ IntranetCmd::IntranetCmd()
 
 IntranetCmdStore::IntranetCmdStore() {}
 
-uint8_t IntranetCmdStore::get_id() const { return data_.id; }
-void IntranetCmdStore::set_id(uint8_t value) { data_.id = value; }
+uint16_t IntranetCmdStore::get_id() const { return data_.id; }
+void IntranetCmdStore::set_id(uint16_t value) { data_.id = value; }
 
 uint8_t IntranetCmdStore::get_value() const { return data_.value; }
 void IntranetCmdStore::set_value(uint8_t value) { data_.value = value; }
